+++ apps/openmw/mwmechanics/character.hpp
@@ -276,7 +276,7 @@ namespace MWMechanics
 
         void playBlendedAnimation(const std::string& groupname, const MWRender::AnimPriority& priority, int blendMask,
             bool autodisable, float speedmult, std::string_view start, std::string_view stop, float startpoint,
-            size_t loops, bool loopfallback = false) const;
+            uint32_t loops, bool loopfallback = false) const;
         bool playGroup(std::string_view groupname, int mode, uint32_t count, bool scripted = false);
         bool playGroupLua(std::string_view groupname, float speed, std::string_view startKey, std::string_view stopKey,
             uint32_t loops, bool forceLoop);
